# src/main/java/dev/simonfischer/profiler/models/exception/reponse/ErrorResponse.java
package dev.simonfischer.profiler.models.exception.reponse;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private int statusCode;
    private String message;
}